==> manuscript/polygenic-scores/code/01h_tables_to_rds.R <==
library(data.table)

saveRDS(fread("../MGRB.phase2.SNPtier12.match.vqsr.minrep.locusannot.WGStier12.unrelated.nocancer.over70.manual_polygenic_scores.GnomAD_NFE_AFs.popstar.externalref.permscores", stringsAsFactors = TRUE, data.table = FALSE), "../MGRB.phase2.SNPtier12.match.vqsr.minrep.locusannot.WGStier12.unrelated.nocancer.over70.manual_polygenic_scores.GnomAD_NFE_AFs.popstar.externalref.permscores.rds")
saveRDS(fread("../MGRB.phase2.SNPtier12.match.vqsr.minrep.locusannot.WGStier12.unrelated.nocancer.over70.manual_polygenic_scores.GnomAD_NFE_AFs.popstar.internalref.permscores", stringsAsFactors = TRUE, data.table = FALSE), "../MGRB.phase2.SNPtier12.match.vqsr.minrep.locusannot.WGStier12.unrelated.nocancer.over70.manual_polygenic_scores.GnomAD_NFE_AFs.popstar.internalref.permscores.rds")
saveRDS(fread("../MGRB.phase2.SNPtier12.match.vqsr.minrep.locusannot.WGStier12.unrelated.nocancer.over70.manual_polygenic_scores.GnomAD_NFE_AFs.popstar.externalref.permsummary", stringsAsFactors = TRUE, data.table = FALSE), "../MGRB.phase2.SNPtier12.match.vqsr.minrep.locusannot.WGStier12.unrelated.nocancer.over70.manual_polygenic_scores.GnomAD_NFE_AFs.popstar.externalref.permsummary.rds")
saveRDS(fread("../MGRB.phase2.SNPtier12.match.vqsr.minrep.locusannot.WGStier12.unrelated.nocancer.over70.manual_polygenic_scores.GnomAD_NFE_AFs.popstar.internalref.permsummary", stringsAsFactors = TRUE, data.table = FALSE), "../MGRB.phase2.SNPtier12.match.vqsr.minrep.locusannot.WGStier12.unrelated.nocancer.over70.manual_polygenic_scores.GnomAD_NFE_AFs.popstar.internalref.permsummary.rds")
saveRDS(fread("../MGRB.phase2.SNPtier12.match.vqsr.minrep.locusannot.WGStier12.unrelated.manual_polygenic_scores.GnomAD_NFE_AFs.popstar.externalref.scores", stringsAsFactors = TRUE, data.table = FALSE), "../MGRB.phase2.SNPtier12.match.vqsr.minrep.locusannot.WGStier12.unrelated.manual_polygenic_scores.GnomAD_NFE_AFs.popstar.externalref.scores.rds")
