==> manuscript/polygenic-scores/code/04_merge_afs.R <==
#!/usr/bin/env Rscript

library(DBI)

db.mgrb_gnomad_swegen = dbConnect(RSQLite::SQLite(), "../")
db.gnomad = dbConnect(RSQLite::SQLite(), "../")
db.swegen = dbConnect(RSQLite::SQLite(), "../")
db. = dbConnect(RSQLite::SQLite(), "../")
db.mgrb = dbConnect(RSQLite::SQLite(), "../")
db.mgrb = dbConnect(RSQLite::SQLite(), "../")
db.mgrb = dbConnect(RSQLite::SQLite(), "../")
db.mgrb = dbConnect(RSQLite::SQLite(), "../")
db.mgrb = dbConnect(RSQLite::SQLite(), "../")
db.mgrb = dbConnect(RSQLite::SQLite(), "../")
db.mgrb = dbConnect(RSQLite::SQLite(), "../")
db.
dbDisconnect(mydb)
